requirements.txt
awscli==1.19.5
apache-airflow-providers-snowflake==1.3.0
pendulum==2.1.2
